requirements.txt
pyyaml
xmltodict
pyinstaller